Moved plotter for ply file edge lengths into ply_to_input.py to help with testing.  Marked that same plotter function as do-not-test.  Finished a proof of concept for setting up sets of tests using both a mocked function an example ply file.
